Update Vertical Pagination Bar
This component is now configurable to be vertical or horizontal
